+++ src/ome_zarr_converters_tools/fractal/__init__.py
@@ -1,5 +1,12 @@
 """API for building OME-Zarr converters tasks for Fractal."""
 
+from ome_zarr_converters_tools.fractal._compound_task_wrapper import (
+    MultiprocessingRunner,
+    RunnerType,
+    SequentialRunner,
+    ThreadedRunner,
+    exec_compound_task,
+)
 from ome_zarr_converters_tools.fractal._compute_task import (
     ImageListUpdateDict,
     generic_compute_task,
@@ -26,10 +33,15 @@ __all__ = [
     "ChannelInfoUI",
     "ConvertParallelInitArgs",
     "ImageListUpdateDict",
+    "MultiprocessingRunner",
     "PixelSizeModel",
+    "RunnerType",
+    "SequentialRunner",
+    "ThreadedRunner",
     "cleanup_if_exists",
     "converters_tools_models",
     "dump_to_json",
+    "exec_compound_task",
     "generic_compute_task",
     "remove_json",
     "setup_images_for_conversion",